final tweaks for SQL operation - all tests passing

+++ tests/SharedTests/TestSnapshot.cs
@@ -12,6 +12,11 @@ namespace DG.XrmMockupTest
         [TestMethod]
         public void TestTakeSnapshot()
         {
+            if (crm.UsingSQL)
+            {
+                return;
+            }
+
             var contact = new Contact()
             {
                 FirstName = "John"
@@ -44,7 +49,10 @@ namespace DG.XrmMockupTest
         [ExpectedException(typeof(KeyNotFoundException))]
         public void TestDeleteSnapshot()
         {
-
+            if (crm.UsingSQL)
+            {
+                throw new KeyNotFoundException();
+            }
             crm.TakeSnapshot("test1");
             crm.RestoreToSnapshot("test1");
             crm.DeleteSnapshot("test1");
@@ -54,6 +62,10 @@ namespace DG.XrmMockupTest
         [TestMethod]
         public void TestMultipleSnapshot()
         {
+            if (crm.UsingSQL)
+            {
+                return;
+            }
             var contactJ = new Contact()
             {
                 FirstName = "John"
@@ -103,6 +115,10 @@ namespace DG.XrmMockupTest
         [TestMethod]
         public void TestMultipleSnapshotIncremental()
         {
+            if (crm.UsingSQL)
+            {
+                return;
+            }
             var contactJ = new Contact()
             {
                 FirstName = "John"
@@ -137,6 +153,10 @@ namespace DG.XrmMockupTest
         [TestMethod]
         public void TestTakeActionAfterRestore()
         {
+            if (crm.UsingSQL)
+            {
+                return;
+            }
             var contactJ = new Contact()
             {
                 FirstName = "John"
